+++ ucare/uploadapi.go
@@ -6,8 +6,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"reflect"
-	"time"
 
 	"github.com/uploadcare/uploadcare-go/v2/internal/config"
 )
@@ -15,13 +13,15 @@ import (
 type uploadAPIClient struct {
 	authFunc UploadAPIAuthFunc
 
-	conn *http.Client
+	conn  *http.Client
+	retry *RetryConfig
 }
 
 func newUploadAPIClient(creds APICreds, conf *Config) Client {
 	c := uploadAPIClient{
 		authFunc: simpleUploadAPIAuthFunc(creds),
 		conn:     conf.HTTPClient,
+		retry:    conf.Retry,
 	}
 
 	if conf.SignBasedAuthentication {
@@ -70,67 +70,81 @@ func (c *uploadAPIClient) Do(
 	req *http.Request,
 	resdata interface{},
 ) error {
-	tries := 0
-try:
-	tries++
+	for tries := 1; ; tries++ {
+		if tries > 1 && req.GetBody != nil {
+			var err error
+			req.Body, err = req.GetBody()
+			if err != nil {
+				return err
+			}
+		}
 
-	if tries > 1 && req.GetBody != nil {
-		var err error
-		req.Body, err = req.GetBody()
+		log.Debugf("making %d request: %s %+v", tries, req.Method, req.URL)
+
+		resp, err := c.conn.Do(req)
 		if err != nil {
 			return err
 		}
-	}
 
-	log.Debugf("making %d request: %s %+v", tries, req.Method, req.URL)
+		retry, err := c.handleResponse(resp, resdata, tries)
+		if err != nil || !retry {
+			return err
+		}
+	}
+}
 
-	resp, err := c.conn.Do(req)
-	if err != nil {
-		return err
-	}
-	if req.Body != nil {
-		defer func() { _ = req.Body.Close() }()
-	}
+func (c *uploadAPIClient) handleResponse(
+	resp *http.Response,
+	resdata interface{},
+	tries int,
+) (bool, error) {
+	defer func() { _ = resp.Body.Close() }()
 
 	log.Debugf("received response: %+v", resp)
 
 	switch resp.StatusCode {
-	case 400, 403:
+	case 400:
 		data, err := io.ReadAll(resp.Body)
 		if err != nil {
-			return err
+			return false, err
 		}
-		_ = resp.Body.Close()
-		switch resp.StatusCode {
-		case 400:
-			return reqValidationErr{respErr{string(data)}}
-		case 403:
-			return reqForbiddenErr{respErr{string(data)}}
+		return false, ValidationError{APIError{StatusCode: 400, Detail: string(data)}}
+	case 403:
+		data, err := io.ReadAll(resp.Body)
+		if err != nil {
+			return false, err
 		}
+		return false, ForbiddenError{APIError{StatusCode: 403, Detail: string(data)}}
 	case 413:
-		return ErrFileTooLarge
+		return false, ErrFileTooLarge
 	case 429:
-		if tries > config.MaxThrottleRetries {
-			return throttleErr{}
-		}
-		// retry after is not returned from the upload API
-		select {
-		case <-req.Context().Done():
-			return req.Context().Err()
-		case <-time.After(5 * time.Second):
-		}
-		goto try
+		return handleThrottle(resp.Request.Context(), resp, c.retry, tries)
 	default:
+		if resp.StatusCode >= 400 {
+			body, err := io.ReadAll(resp.Body)
+			if err != nil {
+				return false, err
+			}
+			var apiErr APIError
+			if json.Unmarshal(body, &apiErr) != nil || apiErr.Detail == "" {
+				detail := string(body)
+				if detail == "" {
+					detail = http.StatusText(resp.StatusCode)
+				}
+				apiErr.Detail = detail
+			}
+			apiErr.StatusCode = resp.StatusCode
+			return false, apiErr
+		}
 	}
 
-	if resdata == nil || reflect.ValueOf(resdata).IsNil() {
-		return nil
+	if isNilResponseData(resdata) {
+		return false, nil
 	}
-	err = json.NewDecoder(resp.Body).Decode(&resdata)
-	if err != nil {
-		return err
-	}
-	_ = resp.Body.Close()
 
-	return nil
+	if err := json.NewDecoder(resp.Body).Decode(resdata); err != nil {
+		return false, err
+	}
+
+	return false, nil
 }